lambda_layers/generateToken/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile '.\requirements.in'
#
boto3==1.20.5
    # via -r .\requirements.in
botocore==1.23.5
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via requests
cryptography==35.0.0
    # via pyjwt
idna==3.3
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via -r .\requirements.in
python-dateutil==2.8.2
    # via botocore
requests==2.26.0
    # via -r .\requirements.in
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.7
    # via
    #   botocore
    #   requests
